Fix deserializes integer double in context
