--- android/beagle/src/main/java/br/com/zup/beagle/android/context/ContextDataValueResolver.kt
@@ -1,31 +1,0 @@
-/*
- * Copyright 2020 ZUP IT SERVICOS EM TECNOLOGIA E INOVACAO SA
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package br.com.zup.beagle.android.context
-
-import org.json.JSONArray
-import org.json.JSONObject
-
-internal class ContextDataValueResolver {
-
-    fun parse(value: Any?): Any? {
-        return when (value) {
-            is Collection<*> -> JSONArray(value)
-            is Map<*, *> -> JSONObject(value)
-            else -> value
-        }
-    }
-}
